version: Move GetKernelVersion
Move the GetKernelVersion() from pkg/datapath/linux to pkg/version so
that it can be included by pkg/maps/lbmap which is included by
pkg/datapath/linux. Otherwise, we have a circular dependency.

Also, make the function to return an error instead of calling log.Fatal,
as it's caller's responsibility to decide what to do with an error.
